requirements.txt
django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.7
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
